refactor: make DummyDA publicly available
